+++ sample-prop.json
@@ -1,5 +1,5 @@
 {
   "token": "",
   "prefix": "",
-  "mongodb-address": ""
+  "mongodb-address": "mongodb://localhost:27017/"
 }
